Fix broken metadata package builder
The logger was broken due to a typo. Specifically, we set the value
of self._project_deps equal to the return of _log_metadata(). This
is always None and therefore was squashing the intended value.

Also shorten self._project_dependencies to self._project_deps and
make logging a bit better.